% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreement_analysis.R
\name{agreement_analysis}
\alias{agreement_analysis}
\title{Agreement Analysis}
\usage{
agreement_analysis(
  data,
  vars = names(data),
  grouping,
  rv,
  crit = 0.7,
  type = "df",
  label,
  n_sim = 10000
)
}
\arguments{
\item{data}{A data-frame}

\item{vars}{Vector of character strings with names of vairables to be analyzed}

\item{grouping}{Grouping variable. Either as a character String or a vector}

\item{rv}{Number of response options for calculating random variance}

\item{crit}{Critical agreement level}

\item{label}{Character vector with new variable labels}

\item{n_sim}{Number of simulations for the agreement analyses}
}
\value{
A Data-frame
}
\description{
Agreement Analysis
}
