% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_CI.R
\name{alpha_CI}
\alias{alpha_CI}
\title{alpha_CI}
\usage{
alpha_CI(alpha, n, items, ci)
}
\arguments{
\item{alpha}{Alpha value}

\item{n}{Sample size}

\item{items}{Number of items}

\item{ci}{Confidence level (e.g. 0.95 for 95 percent)}
}
\value{
a numeric vector containing the upper and lower bounds of the
confidence interval
}
\description{
Calculate the confidence interval for Cronbach's alpha
}
\examples{
alpha_CI(0.8, 10, 50, 0.99)

}
