% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_table.R
\name{alpha_table}
\alias{alpha_table}
\title{Table with alpha values}
\usage{
alpha_table(data, VAR, labels = NULL, round = 2, CI = TRUE,
  conf_level = 0.95, check_key = TRUE, omega = FALSE, key = NULL,
  RMSEA = FALSE)
}
\arguments{
\item{data}{A data Frame}

\item{VAR}{A list containing vectors with variable names. Each list element defines one scale.}

\item{labels}{Label names for scales}

\item{round}{Rounds values to given decimal position}

\item{CI}{If TRUE confidence intervals are calculated}

\item{conf_level}{Confidence level (e.g. 0.95 for 95 percent)}

\item{check_key}{Check_key for the psych::alpha function}

\item{omega}{If TRUE Omega reliability estimation is calculated}

\item{key}{Optional key argument for the psych::alpha function}

\item{RMSEA}{If TRUE RMSEA is calculated}
}
\description{
Returns a list of alpha cronbachs. VAR is a list of character strings
}
