% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_missing.R
\name{analysis_missing}
\alias{analysis_missing}
\title{Analyse missing values}
\usage{
analysis_missing(dat, scale)
}
\arguments{
\item{dat}{A data-frame}

\item{scale}{A list with character vectors. Each list element containing the variable names of a scale}
}
\value{
A data-frame
}
\description{
Analyse missing values
}
