% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_convert_csv.R
\name{batch_convert_csv}
\alias{batch_convert_csv}
\title{Batch convert csv
Convert all csv fies in a folder to xlsx and rds files}
\usage{
batch_convert_csv(folder, xlsx = TRUE, rds = TRUE, move_csv = "",
  dec = ",", sep = ";", ...)
}
\arguments{
\item{folder}{Character string with source folder}

\item{xlsx}{If TRUE creates xlsx files.}

\item{rds}{If TRUE creates rds files.}

\item{move_csv}{If not an empty string it names a folder to move csv to after conversion}

\item{dec}{Decimal sign}

\item{sep}{Separator sign}

\item{...}{Further functions for the underlying read.csv. (e.g. fileEncoding = "WINDOWS-1252")}
}
\value{
Writes xlsx and/or rds files into the source folder for each csv file.
}
\description{
Batch convert csv
Convert all csv fies in a folder to xlsx and rds files
}
