% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Mean center a variable}
\usage{
center(x, min_valid, max_na, grouping)
}
\arguments{
\item{x}{A vector of numeric values to be mean centered.}

\item{min_valid}{Minimal number of valid values that is required for
calculation. A value between 0 and 1 indicates a proportion of values
(e.g., 0.5 = 50 percent of values have to be valid).}

\item{max_na}{Maximum number of NAs that are allowed before returning NA. A
value between 0 and 1 indicates a proportion of values (e.g., 0.5 = 50
percent NAs are allowed).}

\item{grouping}{A variable with group values for calculating grouped
centers.}
}
\value{
A vector with mean centered values
}
\description{
This function mean centers a variable, optionally within groups. It also
allows to set minimal valid values or maximal allowed NAs for the calculation.
If the conditions are not met, NA values are returned.
}
\details{
If both \code{min_valid} and \code{max_na} are provided, both conditions
have to be met for the calculation to proceed.
If \code{grouping} is provided, the mean centering is done
within the groups defined by \code{grouping}.
Missing values (NAs) are ignored in the calculation of means.
If the conditions for \code{min_valid} or \code{max_na} are not met,
NA values are returned for all elements of \code{x}.
}
\examples{
center(c(1:10, NA, NaN))
center(mtcars$mpg, grouping = mtcars$cyl)
center(c(1:5, NA, NA), min_valid = 4)
center(c(1:5, NA, NA), max_na = 1)
center(c(1:5, NA, NA), min_valid = 0.7)
center(c(1:5, NA, NA), max_na = 0.3)
center(c(1:5, NA, NA), min_valid = 0.7, max_na = 1)
center(c(1:5, NA, NA), min_valid = 4, max_na = 0)

}
\author{
Juergen Wilbert
}
