% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{check_args}
\alias{check_args}
\title{Checks arguments and returns a single error message for all errors.}
\usage{
check_args(...)
}
\arguments{
\item{...}{}
}
\description{
Checks arguments and returns a single error message for all errors.
}
