% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_scale.R
\name{check_scale}
\alias{check_scale}
\title{Check scale for typos}
\usage{
check_scale(data, var.names, min, max)
}
\arguments{
\item{data}{A data-frame}

\item{var.names}{Character vector with variable names}

\item{min}{minimum of scale}

\item{max}{maximum of scale}
}
\value{
A dataframe containing all cases with valies above and below the alowed values.
}
\description{
Check scale for typos
}
