% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIfromSD.R
\name{CIfromSD}
\alias{CIfromSD}
\title{Calculate confidence interval from standard deviation}
\usage{
CIfromSD(mean, sd, n, alpha = 0.95)
}
\arguments{
\item{mean}{the mean of the sample}

\item{sd}{the standard deviation of the sample}

\item{n}{the number of samples}

\item{alpha}{the % confidence interval to be calculated, defaults to 0.95}
}
\value{
returns the lower and upper boundaries of the confidence interval
}
\description{
Calculate confidence interval from standard deviation
}
