% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_row.R
\name{combine_data_frames}
\alias{combine_data_frames}
\title{Combine multiple data frames into a single data frame}
\usage{
combine_data_frames(...)
}
\arguments{
\item{...}{The data frames to combine}
}
\value{
A combined data frame
}
\description{
This function takes in multiple data frames and combines them into a single data frame. Missing columns and rows between data frames are filled with `NA` values.
}
\examples{
a <- data.frame(id = 1:3, name = c("Alice", "Bob", "Charlie"))
b <- data.frame(id = 4:6, age = c(25, 30, 35))
combine_data_frames(a,b)
# id    name age
# 1  Alice  NA
# 2    Bob  NA
# 3 Charlie  NA
# 4     NA  25
# 5     NA  30
# 6     NA  35
}
