% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{conf_int}
\alias{conf_int}
\title{Calculate Confidence Interval}
\usage{
conf_int(x, ci = 0.95, r = NA, s = NA, se = NA, n = NA)
}
\arguments{
\item{x}{A sample mean}

\item{ci}{Numeric, confidence level (default = 0.95)}

\item{r}{Numeric, correlation coefficient (default = NA)}

\item{s}{Numeric, standard deviation of population (default = NA)}

\item{se}{Numeric, standard error of the sample mean (default = s * sqrt(1 -
r))}

\item{n}{Numeric, sample size (default = NA)}
}
\value{
A list containing confidence intervals from Z and t methods.
}
\description{
Calculate the confidence interval of a given sample mean.
}
\examples{
conf_int(x = 12, ci = 0.99, s = 2.5, n = 24)
conf_int(x = 12, s = 2.5, r = 0.6)
}
