% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blogsite.R
\name{create_blogsite}
\alias{create_blogsite}
\title{Create a blogsite for Rmakrdown webpages}
\usage{
create_blogsite(
  source_dir = "posts",
  template = "_blog.Rmd",
  outfile = "blog.Rmd",
  subdir = "posts",
  render = FALSE,
  date_format = "\%A, \%B \%d",
  filepattern = c(".rmd", ".md", ".Rmd"),
  no_temp_files = TRUE,
  root = getwd()
)
}
\arguments{
\item{source_dir}{A directory path with blogposts}

\item{template}{A template file to create a blog-website from}

\item{outfile}{Output Rmd file with list of blog entries and references (based on template)}

\item{subdir}{Subdirectory where blog entries will be saved}

\item{render}{Default is FALSE. When TRUE, all Markdown files in subdirectory are rendered to html. Otherwise, the function just takes available html files in the subdirectory folder}

\item{date_format}{A character string how to format the Date}

\item{filepattern}{File extensions that are included.}

\item{no_temp_files}{Logical. If TRUE, files (Markdown and html) starting with "_" are exluded}

\item{root}{Root directory of the project.}
}
\value{
A markdown file in the project directory.
}
\description{
Just a little helper for something like Blogs with Rmarkdown pages
}
