% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportFig.R
\name{exportFig}
\alias{exportFig}
\title{Exporting a graphic figure}
\usage{
exportFig(
  figure,
  filename = "figure",
  widthpx = 1920,
  heightpx = 1080,
  res = 200,
  format = c("png", "pdf"),
  dir = "output",
  prefix = "date",
  replace = F
)
}
\arguments{
\item{figure}{the graphics to be exported.}

\item{filename}{a string which the name of the figure to be exported}

\item{widthpx}{an integer which specifies the width of the exported figure in pixels}

\item{heightpx}{an integer which specifies the height of the exported figure in pixels}

\item{res}{an integer which specifies the resolution of the exported figure for png.}

\item{format}{a character vector which specifies the exported format(s). Currently supports "png" and "pdf".}

\item{dir}{a character string which specifies the directory that the figure will be exported, relative to
the current working directory. If the directory does not exist, it will be created.}

\item{prefix}{a character string which specifies the time of prefix. "date" specifies the current date, "time" prefixes
both the date and time to the filename, and NULL removes the prefix altogether.}

\item{replace}{a logical value. If set to TRUE,  files with the same file name will be overwritten on export. If set to FALSE,
the software will check for duplicate filenames and append a padded integer to the filename to avoid overwriting.}
}
\value{

}
\description{
This function provides a simple interface to export figures generated in R to multiple formats.
}
