% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportFile.R
\name{exportFile}
\alias{exportFile}
\title{Export data.frame or matrix objects}
\usage{
exportFile(
  object,
  filename = "table",
  prefix = "date",
  file_format = c("tsv", "csv", "xlsx"),
  dir = "output/",
  replace = F
)
}
\arguments{
\item{object}{a single 'data.frame' or 'matrix' object that is to be exported.}

\item{filename}{a character vector of the intended file name of the object to be exported.
Defaults to "table".}

\item{prefix}{a character vector containing one of 'date', 'time', or NULL. 'date'
prefixes date (e.g. 2021-09-25), whereas 'time' prefixes both time and date. NULL will
remove the prefix altogether.}

\item{file_format}{a character vector containing one or more of "tsv", "csv", and "xlsx".
This species the format(s) of the exported file(s).}

\item{dir}{a character vector which specifies the location of the exported files relative
to the working directory}

\item{replace}{a logical value specifying whether the function should check for redundant
file names. If replace is set to FALSE, a new filename will be iteratively generated with
a numeric suffix until a unique file name is reached.}
}
\description{
Export a data.frame or matrix into several common formats, and automatically prefixes a
date/time field. Also optionally renames exported files to prevent replacing files with
the same name.
}
\examples{
\dontrun{
exportFile(x,
           filename = "data.frame",
           prefix = "time",
           file_format = c("tsv", "xlsx"),
           dir = "tables",
           replace = T)
}
}
