% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_l2_data.R
\name{fill_missing_l2}
\alias{fill_missing_l2}
\title{Fills in missing data for wide format}
\usage{
fill_missing_l2(data, id, vars)
}
\arguments{
\item{data}{A data frame}

\item{id}{Character string with L2 variable}

\item{vars}{Vector of strings with varible names}
}
\value{
A data frame with replaced values
}
\description{
Fills in missing data for wide format
}
\examples{
x <- data.frame(id = rep(1:3, each = 3), gender = c(1, 1, NA, 0, 0, 0, 1, NA, NA))
fill_missing_l2(x, "id", "gender")

x <- data.frame(id = rep(1:3, each = 3), gender = c(1, 1, NA, 0, 0, 0, 1, 0, NA))
fill_missing_l2(x, "id", "gender")
}
