% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_z.R
\name{fisher_z}
\alias{fisher_z}
\alias{inv_fisher_z}
\title{Fisher's Z Transformation}
\usage{
fisher_z(r)

inv_fisher_z(z)
}
\arguments{
\item{r}{The correlation coefficient to be transformed.}

\item{z}{The Fisher's Z transformed value.}
}
\value{
The Fisher's Z transformed value.

The original correlation coefficient.
}
\description{
The Fisher's Z transformation is a common method for stabilizing the variance
of the sample correlation coefficient.

The inverse Fisher's Z transformation is used to convert a Fisher's Z value
back to a correlation coefficient.
}
\details{
The Fisher's Z transformation is defined as: \code{0.5 * log((1 + r)
/ (1 - r))}

The inverse Fisher's Z transformation is defined as: \code{(exp(2 *
z) - 1) / (exp(2 * z) + 1)}
}
\examples{
# Example usage:
r <- 0.8
fisher_z(r)

# Example usage:
z <- 0.972
inv_fisher_z(z)

}
\references{
Fisher, R. A. (1915). Frequency distribution of the values of the
correlation coefficient in samples from an indefinitely large population.
\emph{Biometrika}, 10(4), 507-521.
}
\seealso{
\code{\link{cor}}, \code{\link{cor.test}},
\code{\link{inv_fisher_z}}

\code{\link{cor}}, \code{\link{cor.test}}, \code{\link{fisher_z}}
}
\author{
Your Name
}
