% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{flip}
\alias{flip}
\title{Flip a Data Frame or Matrix}
\usage{
flip(x, rownames = NULL)
}
\arguments{
\item{x}{A data frame or matrix to be flipped.}

\item{rownames}{Logical or string. If `TRUE`, row names will be included as the 
first column. If `character`, the column will be named respectively.}
}
\value{
A data frame with flipped rows and columns. If \code{rownames} is 
  \code{TRUE}, row names (if `x` is a matrix) or the first column will 
  contain the original row names (if `x` is a data frame).
}
\description{
This function takes a data frame or matrix and flips its rows and columns,
effectively transposing the data. Optionally, it can include row names in the 
resulting data frame.
}
\examples{
# Create a sample data frame
df <- data.frame(A = 1:3, B = 4:6, C = 7:9)

# Flip the data frame
flipped_df <- flip(df)

# Flip the data frame with row names included
flipped_df_with_rownames <- flip(df, rownames = TRUE)

}
\seealso{
\code{\link{t}} for transposing matrices and data frames.
}
