% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean.R
\name{geomean}
\alias{geomean}
\alias{rowGeomean}
\alias{colGeomean}
\title{Vectorized calculation of geometric mean}
\usage{
geomean(x, na.rm = T)

rowGeomean(lx, na.rm = T)

colGeomean(lx, na.rm = T)
}
\arguments{
\item{x}{a numeric vector. Geometric mean can only be calculated for positive numbers.}

\item{na.rm}{If TRUE, any missing values are ignored, otherwise not.}
}
\description{
Geometric mean is a robust measure of central tendency. This function introduces a
geometric mean calculation and to calculate the geometric mean for rows and
columns of matrix by wrapping functions from the package matrixStats.
}
