% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attribute.R
\name{get_wmisc_attributes}
\alias{get_wmisc_attributes}
\alias{set_wmisc_attributes}
\title{Get or Set 'wmisc' Attributes}
\usage{
get_wmisc_attributes(x)

set_wmisc_attributes(x, ...)
}
\arguments{
\item{x}{An R object from which to get or to which to set attributes.}

\item{...}{Named attributes to set for the object.
#' @return
For \code{get_wmisc_attributes}, a list of attributes stored under "wmisc".
For \code{set_wmisc_attributes}, the original object with updated "wmisc" attributes}
}
\description{
These functions allow you to get or set custom attributes under the "wmisc" namespace for any R object.
}
\keyword{internal}
