% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_attribute.R
\name{has_attribute}
\alias{has_attribute}
\title{Checks for attributes}
\usage{
has_attribute(.x, expr, dic = FALSE)
}
\arguments{
\item{.x}{A dataframe, a list or a single object. If a dataframe is provided,
the function will check the attributes of each column. If a list is
provided, the function will check the attributes of each element. If a
single object is provided, the function will check the attributes of that
object.}

\item{expr}{A logical expression based on the attributes of the object(s).
The expression should be written in terms of the attributes of the object(s).
For example, if you want to check for an attribute called "my_attr" that
should be equal to "my_value", you would write the expression as \code{my_attr == "my_value"}.}

\item{dic}{A logical value indicating whether to look directly and only
\code{scaledic} attributes.}
}
\value{
A logical vector indicating which columns (or elements) have the
specified attribute.
}
\description{
This function takes an object returns a logical vector of the same length as
the number of columns in the object, where each element is TRUE if the
corresponding column has the specified attribute values and FALSE otherwise.
}
