% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_class.R
\name{list_class}
\alias{list_class}
\title{Show class of all cols in a data.frame}
\usage{
list_class(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
A data.frame with variable names and classes
}
\description{
Show class of all cols in a data.frame
}
