% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit-prob.R
\name{logit2prob}
\alias{logit2prob}
\alias{prob2logit}
\title{Convert Logit to Probability and vice versa}
\usage{
logit2prob(logit)

prob2logit(prob)
}
\arguments{
\item{logit}{Numeric. The logit value to be converted to probability.}

\item{prob}{Numeric. A probability to be converted to a logit.}
}
\value{
Numeric.
}
\description{
This function takes a logit value and converts it to a probability using the logistic function and vice versa.
}
\examples{
logit2prob(0) # Returns 0.5
logit2prob(2) # Returns 0.8807971
prob2logit(0.8807971)
}
