% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixCenter.R
\name{matrixCenter}
\alias{matrixCenter}
\title{Centering a numeric matrix}
\usage{
matrixCenter(lx, MARGIN = 1, scale = F)
}
\arguments{
\item{lx}{a numeric matrix}

\item{MARGIN}{a vector giving the subscripts which the function will be applied over. E.g.,
for a matrix 1 indicates rows, 2 indicates columns, c(1, 2) indicates rows and columns.}

\item{scale}{either a logical value or a numeric-alike vector of length equal to the number
of columns of x.}
}
\value{

}
\description{
Plotting heatmaps of gene expression data in numeric matrices often is the most clear when
each gene is centered on 0. Since most expression data is log2 transformed, scaling of data
is not desirable.
}
