% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_corr.R
\name{mean_corr}
\alias{mean_corr}
\title{Average correlations by Fisher Z transformation}
\usage{
mean_corr(r)
}
\arguments{
\item{r}{A vector of correlation coefficients}
}
\value{
The average correlation
}
\description{
Average correlations by Fisher Z transformation
}
\examples{
rs <- c(0.2, 0.5, 0.9)
mean(rs)
mean_corr(rs)
}
