% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_multilevel.R
\name{merge_multilevel}
\alias{merge_multilevel}
\title{Helps merging multilevel datasets}
\usage{
merge_multilevel(dat.l1, dat.l2 = NULL, id, var.mean)
}
\arguments{
\item{dat.l1}{A data Frame for Level 1}

\item{dat.l2}{A data frame for Level 2}

\item{id}{Stirng with variable name for matching datasets}

\item{var.mean}{Vector with strings of L1 variables that are aggregated as Level2 means}
}
\value{
A data frame
}
\description{
Helps merging multilevel datasets
}
