% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_average.R
\name{moving_average}
\alias{moving_average}
\title{Moving Average}
\usage{
moving_average(x, xLag, FUN = mean)
}
\arguments{
\item{x}{Vector}

\item{xLag}{Lag for calculating the average}

\item{FUN}{Function for calculating the average (e.g., FUN = mean)}
}
\description{
Moving Average
}
