% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_blog_entry.R
\name{new_blog_entry}
\alias{new_blog_entry}
\title{Create blog entry}
\usage{
new_blog_entry(
  title = "New Blog Entry",
  author = "Juergen Wilbert",
  date = format(Sys.time(), "\%Y-\%m-\%d"),
  abstract = "",
  keywords = "blog",
  subdir = "posts",
  ...
)
}
\arguments{
\item{title}{Title}

\item{author}{Author}

\item{date}{Date}

\item{abstract}{Abstract}

\item{keywords}{Keywords}

\item{subdir}{Directory with posts}

\item{...}{Further YAML parameters}
}
\value{
Nothing
}
\description{
Create blog entry
}
