% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_contingency_table.R
\name{nice_contingency_table}
\alias{nice_contingency_table}
\title{Create a Contingency Table with One or Two Summary Functions}
\usage{
nice_contingency_table(
  rows,
  columns,
  var,
  fn = list(Mean = mean, n = function(.) length(.)),
  auto_labels = TRUE,
  var_label = NULL,
  rows_label = NULL,
  columns_label = NULL,
  fn_label = NULL,
  title = NULL,
  footnote = NULL,
  file = NULL,
  remove_missing = TRUE,
  sort = TRUE,
  ...
)
}
\arguments{
\item{rows}{A vector or factor for the row grouping variable. If \code{columns}
and \code{var} are missing, \code{rows} can be a list where its first element
is the row variable, second is the column variable, and third is the variable
of interest.}

\item{columns}{A vector or factor for the column grouping variable. If missing,
it will be extracted from the second element of \code{rows} (if \code{rows}
is a list).}

\item{var}{A numeric vector containing the values to be summarized. If missing,
it will be extracted from the third element of \code{rows} (if \code{rows}
is a list).}

\item{fn}{A named list of functions to apply to \code{var} for each combination
of \code{rows} and \code{columns}. Defaults to
\code{list(Mean = mean, n = \(.) length(.))}. If only one function is
provided, the table will display that result. If two are provided, the results
will be concatenated (e.g., \code{"Mean (n)"}).}

\item{auto_labels}{A logical value indicating whether to automatically retrieve
labels from attributes of \code{rows}, \code{columns}, or \code{var}.
Defaults to \code{TRUE}.}

\item{var_label}{An optional character string specifying the label for
\code{var}.}

\item{rows_label}{An optional character string specifying the label for
\code{rows}.}

\item{columns_label}{An optional character string specifying the label for
\code{columns}.}

\item{fn_label}{An optional character vector specifying the labels for
the functions listed in \code{fn}. If not provided, defaults to the names
of \code{fn}.}

\item{title}{An optional character string specifying the title of the table.
By default, it is constructed from \code{fn_label} and \code{var_label}.}

\item{footnote}{An optional character string specifying a footnote for the
table.}

\item{file}{An optional path to a file where the table output should be saved.}

\item{remove_missing}{A logical value indicating whether rows with missing
values (\code{NA}) in \code{var} should be removed. Defaults to \code{TRUE}.}

\item{sort}{A logical value indicating whether to sort the unique levels of
\code{rows} and \code{columns}. Defaults to \code{TRUE}.}

\item{...}{Additional arguments passed to \code{\link{nice_table}}.}
}
\value{
An html table.
}
\description{
This function creates a contingency table by grouping the data into unique
levels of \code{rows} and \code{columns}, then applies one or more summary
functions (by default \code{mean} and a count via \code{length}) to summarize
the variable of interest (\code{var}). Labels can be assigned manually or
automatically retrieved from variable attributes if available.
}
\details{
If \code{rows}, \code{columns}, and \code{var} are provided separately, the
function will group the data by \code{rows} and \code{columns}, then apply the
functions in \code{fn} to \code{var} within each combination of row and column
levels.

If \code{columns} and \code{var} are not specified, you can instead provide a
list to \code{rows} whose elements are (1) the row variable, (2) the column
variable, and (3) the variable to be summarized.

If two functions are provided via \code{fn}, the resulting table will show
the first function's value followed by the second function's value in
parentheses. For example, if the first function is \code{mean} and the second
is a count function, a cell might display \code{"3.45 (10)"}.

The function attempts to automatically retrieve labels from the data if
\code{auto_labels = TRUE} and the variables have \code{"label"} attributes. If
labels are not found or if \code{var_label}, \code{rows_label}, or
\code{columns_label} is explicitly specified, those provided labels will be
used.

Any \code{NA} values in \code{var} are removed before performing calculations
if \code{remove_missing = TRUE}.

By default, the function also sorts the levels of the row and column variables.
}
\examples{
nice_contingency_table(wmisc:::mtcars_labeled[, c("cyl", "carb", "mpg")])

nice_contingency_table(
  wmisc:::mtcars_labeled[, c("cyl", "am", "mpg")],
  fn = list(Median = median, "mean average deviation" = mad),
  label_na = "-",
)

}
