% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_corrmatrix.R
\name{nice_corrmatrix}
\alias{nice_corrmatrix}
\title{Nice correlation matrix}
\usage{
nice_corrmatrix(
  cr,
  group = NULL,
  upper = FALSE,
  lower = TRUE,
  digits = 2,
  show_r = TRUE,
  show_p = FALSE,
  show_stars = TRUE,
  show_ci = FALSE,
  numbered_columns = TRUE,
  show_descriptives = TRUE,
  nsig_p = 0.1,
  char_nsig,
  char_autocor = "－",
  char_p10 = "✞",
  char_NA = "",
  string_ci = "{break_sign}[{ci_lower},{ci_upper}]",
  string_p = "{break_sign}(p {nice_p(p, equal_sign = TRUE)})",
  caption = NULL,
  title = caption,
  drop_zero = TRUE,
  type = "html",
  file = NULL,
  ...
)
}
\arguments{
\item{cr}{A data frame}

\item{group}{a vector with a grouping variable or a character with a variable
name. If included, calculates multilevel correlations.}

\item{upper}{TRUE if upper triangle should be included.}

\item{lower}{TRUE if lower triangle should be included.}

\item{digits}{Round to given digit position.}

\item{show_r}{TRUE if r-values should be included.}

\item{show_p}{TRUE if r-values should be included.}

\item{show_stars}{TRUE if stars should be included.}

\item{show_ci}{If TRUE, confidence intervals are added.}

\item{numbered_columns}{If TRUE, columns are numbered.}

\item{show_descriptives}{If TRUE, mean and sd columns are added.}

\item{nsig_p}{p level below which correlations are considered not
significant.}

\item{char_nsig}{Character indexing non-significant values.}

\item{char_autocor}{Character for diagonal (e.g. "-" or "1.00").}

\item{char_p10}{Character indexing .10 significance level.}

\item{char_NA}{Character for NA values.}

\item{string_ci}{Character string in glue format for confindence internvals.}

\item{string_p}{Character string in glue format for p-values.}

\item{drop_zero}{If TRUE, leadning zeros are dropped.}

\item{type}{Character string. "df" for data-frame. "html" for html table.}

\item{file}{If TRUE or a filename is provided, a file is exportet (format is
defined by file ending eith html or docx).}

\item{...}{Further arguments passed to the \code{\link[=cor.test]{cor.test()}} function.}

\item{caption/title}{Title for an html table.}
}
\value{
A data-frame or a html table object
}
\description{
From a data.frame object it gives a correlation matrix formatted for overview
}
\examples{
nice_corrmatrix(mtcars)
nice_corrmatrix(mtcars,
  show_p = TRUE,
  show_ci = TRUE,
  show_stars = FALSE,
  show_descriptives = FALSE,
  conf.level = 0.99
)
nice_corrmatrix(mtcars, group = "cyl")
}
