% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_descriptives.R
\name{nice_descriptives}
\alias{nice_descriptives}
\title{Table with descriptive statistics}
\usage{
nice_descriptives(
  data,
  use_col_labels = TRUE,
  title = "Descriptive statistics",
  footnote = NULL,
  file = NULL,
  show_scale = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame}

\item{use_col_labels}{If TRUE, variable names are taken from a label
attribute.}

\item{title}{Title for the table.}

\item{footnote}{Footnote for the table.}

\item{file}{If a file name is provided, the table is saved to this file.}

\item{show_scale}{If TRUE, a column with the scale range is added if the
variable has value labels (either from haven or scaledic attributes).}

\item{...}{Further arguments passed to \code{\link[=nice_table]{nice_table()}}.}
}
\value{
A data frame with descriptive statistics
}
\description{
Table with descriptive statistics
}
\examples{
nice_descriptives(mtcars)

nice_descriptives(mtcars_labeled, auto_labels = TRUE, round = 2)
}
