% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_frequencies.R
\name{nice_frequencies}
\alias{nice_frequencies}
\title{Table with frequency statistics}
\usage{
nice_frequencies(
  data,
  grouping = NULL,
  label = NULL,
  label_grouping = NULL,
  show_missing = TRUE,
  show_percent = TRUE,
  show_percent_cumulative = FALSE,
  percent_base = "column",
  show_total_col = TRUE,
  show_total_row = TRUE,
  auto_labels = TRUE,
  title = NULL,
  footnote = NULL,
  file = NULL,
  round = 1,
  ...
)
}
\arguments{
\item{data}{Vector.}

\item{grouping}{A grouping variable as a vector.}

\item{label}{Set label for the variable name.}

\item{label_grouping}{Set label for the grouping variable name.}

\item{show_missing}{If TRUE, adds a row for the number of missing values.}

\item{show_percent}{If TRUE, adds a column for percentages.}

\item{show_percent_cumulative}{If TRUE, adds a column for cumulative
percentages (only if no grouping variable is provided).}

\item{percent_base}{If show_percent is TRUE, this argument specifies the base
for percentage calculations. Options are "column" (percentages calculated
within each column), "row" (percentages calculated within each row), or
"total" (percentages calculated based on the total count). Default is
"column".}

\item{show_total_col}{If TRUE, adds a column with the total counts and
percentages (if show_percent is TRUE).}

\item{show_total_row}{If TRUE, adds a row with the total counts and
percentages (if show_percent is TRUE).}

\item{auto_labels}{If TRUE, variable names are taken from a label attribute.}

\item{title}{Table title.}

\item{footnote}{Table footnote.}

\item{file}{Filename.}

\item{round}{Number of decimal places for percentages.}

\item{...}{Further arguments passed to \code{\link[=nice_table]{nice_table()}}.}
}
\value{
An html table with frequencies
}
\description{
This function creates a frequency table for a given variable, optionally
grouped by another variable. It can include counts, percentages, and totals,
with customizable labels and formatting options.
}
\details{
The resulting data frame contains the frequency counts and, if specified, the
percentages for each level of the variable and grouping variable. The table can
also include totals for rows and columns, and a row for missing values if desired.
The resulting data frame is decorated with attributes for use with the
\code{wmisc::nice_table()} function, which is called internally to create the final
output.
}
\examples{
nice_frequencies(mtcars_labeled[[11]])

## cross table
nice_frequencies(
  mtcars_labeled$cyl,
  mtcars_labeled$am
)

## barebone table
nice_frequencies(
 mtcars_labeled$cyl,
 mtcars_labeled$am,
 show_missing = FALSE,
 show_percent = FALSE,
 show_total_col = FALSE,
 show_total_row = FALSE
)
nice_frequencies(
  mtcars_labeled$cyl,
  mtcars_labeled$am,
  label = "Cylinders",
  label_grouping = "Transmission",
  title = "Cylinders by Transmission Type"
)
}
