% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_missings_table.R
\name{nice_missings_table}
\alias{nice_missings_table}
\title{Analyze Missing Values in Scales}
\usage{
nice_missings_table(dat, scale, ...)
}
\arguments{
\item{dat}{A data frame containing the data to be analyzed.}

\item{scale}{A named list of character vectors. Each list element defines a scale
by specifying variable names (column names in \code{dat}) belonging to that scale.}

\item{...}{Further arguments passed to the \code{nice_table()} function for formatting the output.}
}
\value{
A data frame summarizing the missing value statistics for each scale with the following columns:
\describe{
\item{Missing}{Number and percentage of missing values across variables in the scale.}
\item{Complete cases}{Number and percentage of cases with no missing value in the scale.}
\item{All missing cases}{Number and percentage of cases where all scale variables are missing.}
}
}
\description{
Summarizes missing values across user-defined scales in a data frame.
}
\details{
This function is useful for assessing the extent of missing data within specific scales
of a dataset. It provides insights into both the overall missings and the distribution
of missing values across cases for each defined scale.
}
\examples{
dat <- data.frame(
  scale1_var1 = c(1, 2, NA, 4,NA),
  scale1_var2 = c(NA, 2, NA, 4,NA),
  scale2_var1 = c(1, NA, 3, 4, 1),
  scale2_var2 = c(NA, NA, NA, 4, 1)
)

scales <- list(
  scale1 = c("scale1_var1", "scale1_var2"),
  scale2 = c("scale2_var1", "scale2_var2")
)

nice_missings_table(dat, scales)

}
\seealso{
\code{\link[=is.na]{is.na()}}, \code{\link[=complete.cases]{complete.cases()}}
}
\author{
Jürgen Wilbert
}
