% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_sem.R
\name{nice_sem}
\alias{nice_sem}
\title{Nice html table for lavaan structure equation model fitted objects}
\usage{
nice_sem(
  x,
  standardized = TRUE,
  show_fitmeasures = c(CFI = "cfi", TLI = "tli", RMSEA = "rmsea", `RMSEA ci lower` =
    "rmsea.ci.lower", SRMR = "srmr", AIC = "aic", BIC = "bic"),
  remove_cols = NULL,
  show_ci = TRUE,
  round = 2,
  ...
)
}
\arguments{
\item{x}{An object returned from the lavaan sem or cfa function.}

\item{standardized}{If TRUE, standardized estimates are shown.}

\item{show_fitmeasures}{A named vector with fit measures.}

\item{remove_cols}{Either column number or column names to be removed.}

\item{show_ci}{If TRUE, adds columns with 95\% confidence intervals.}

\item{round}{Number of digits to round numeric values.}

\item{...}{Further arguments passed to the \code{nice_table()} function.}
}
\value{
An HTML table summarizing the structure equation model.
}
\description{
This function takes an object returned from the \code{lavaan} package's \code{sem} or \code{cfa} functions and produces a nicely formatted HTML table summarizing the model parameters and fit measures. It allows customization of displayed columns, inclusion of confidence intervals, and choice between standardized and unstandardized estimates.
}
\details{
The function organizes the output into sections for latent variables, regressions, covariances
, variances, and intercepts. It also appends key fit measures at the end of the table.
Users can specify which columns to remove, whether to show confidence intervals, and whether to display standardized estimates.
It leverages the \code{nice_table()} function for final formatting, allowing further customization through additional arguments.

The resulting table includes a title and footnote that provide context about the estimation method, optimization algorithm, handling of missing data, and citation information for the \code{lavaan} package.
}
\examples{
nice_sem(wmisc:::lavaan_fit)
nice_sem(wmisc:::lavaan_fit, standardized = FALSE, show_ci = FALSE)
nice_sem(
 wmisc:::lavaan_fit,
 show_fitmeasures = c(
  FI = "cfi", TLI = "tli", RMSEA = "rmsea",
  "SRMR" = "srmr"
 ),
 remove_cols = c("se", "z")
)
 
}
\author{
Juergen Wilbert
}
