% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentage_bar.R
\name{percentage_bar}
\alias{percentage_bar}
\title{Create a Percentage Bar Chart}
\usage{
percentage_bar(perc, label)
}
\arguments{
\item{perc}{The percentage value to be represented as a bar in the chart.}

\item{label}{An optional label to be displayed along with the percentage bar.}
}
\value{
A ggplot2 bar chart representing the given percentage value.
}
\description{
This function creates a percentage bar chart using ggplot2, where the user can specify
the percentage to be represented as a bar along with an optional label.
}
\examples{
# Create a percentage bar chart with 30\% labeled as "Completed"
percentage_bar(30, "Completed")

}
