% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-to-colors.R
\name{plot_discrete_legend}
\alias{plot_discrete_legend}
\title{Plot a discrete color legend for the discrete color scheme.}
\usage{
plot_discrete_legend(values, colors, title = "Values")
}
\arguments{
\item{values}{Numeric or factor vector of values to map (same type as \code{x}).}

\item{colors}{Character vector of hex color codes (same length as \code{values}
and in the same order).}

\item{title}{Title for the legend.}
}
\value{
A plot of the color legend.
}
\description{
Plot a discrete color legend for the discrete color scheme.
}
