% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotalphabeta.R
\name{plotalphabeta}
\alias{plotalphabeta}
\title{Plot ALpha / Beta graph
enerates a graphic for visualizing alpha and beta error}
\usage{
plotalphabeta(m = 0, sd = 1, d = 3, alpha = 0.01, svg.out = "",
  fill = TRUE)
}
\arguments{
\item{m}{Mean}

\item{sd}{Standard Deviation}

\item{d}{Effect size}

\item{alpha}{Alpha Level}

\item{svg.out}{If set a svg file with the given name is created}

\item{fill}{If TRUE intersections are filled out in grey}
}
\description{
Plot ALpha / Beta graph
enerates a graphic for visualizing alpha and beta error
}
