% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primp.R
\name{primp}
\alias{primp}
\title{primp}
\usage{
primp(x, cat = TRUE)
}
\arguments{
\item{x}{A text containing possible styles in curly brackets, separated by
comma, semicolon or space.}

\item{cat}{TRUE or FALSE}
}
\value{
Cat a string or returns string
}
\description{
Get styled console text output
}
\details{
Possible styles are: "bold", "faint", "italic", "underline", "swap",
"strike", "black", "red", "green", "yellow", "blue", "magenta", "cyan",
"white", "lightblack", "lightred", "lightgreen", "lightyellow", "lightblue",
"lightmagenta", "lightcyan", "lightwhite", "bg-black", "bg-red", "bg-green",
"bg-yellow", "bg-blue", "bg-magenta", "bg-cyan", "bg-white", "bg-lightblack",
"bg-lightred", "bg-lightgreen", "bg-lightyellow", "bg-lightblue",
"bg-lightmagenta", "bg-lightcyan", "bg-lightwhite"
}
\examples{
primp("{green bold}Jo Lady, {blue,italic}that's me, {bg-red}in a tree!")
}
