% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_package_version.R
\name{reference_package_version}
\alias{reference_package_version}
\title{Reference package version for markdown}
\usage{
reference_package_version(packages, ...)
}
\arguments{
\item{packages}{Vector of strings with package names.}

\item{...}{Package names}
}
\value{
A markdown formatted for citation of the packages.
}
\description{
Reference package version for markdown
}
