% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{add_label}
\alias{add_label}
\alias{rename_from_labels}
\alias{get_labels}
\title{Various functions that add and retrieve labels for variabes}
\usage{
add_label(x, value)

rename_from_labels(
  data,
  keep = FALSE,
  pattern = "{name}: {label}",
  max_char = NULL
)

get_labels(data, keep = FALSE, pattern = "{name}: {label}", max_char = NULL)
}
\arguments{
\item{x}{A data frame or a vector.}

\item{value}{Either a single character string when \code{x} is a vector or a named
list (\code{columnname = label}) with character strings when \code{x} is a
data.frame.}

\item{data}{A data.frame.}

\item{keep}{If TRUE, new names are a combination of var names and label
names.}

\item{pattern}{Pattern for gleu function when \code{keep = TRUE}.}

\item{max_char}{Maximum number of characters for the label.}
}
\value{
For add_label: A vector or a data.frame with added label
attribute(s).

For rename_from_labels: A data.frame with renamed labels or for
get_labels a character vector with label names.

For \code{get_labels}: A names character vector with labels (and columnnames
as names).
}
\description{
These functions work with an attribute \code{label}. Which has been
introduced by other packages (e.g. \code{haven}) and is implemented into the
R-Studiio environment.
}
