% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_at_percentile.R
\name{split_at_percentile}
\alias{split_at_percentile}
\title{Factorizes a vector by percentiles}
\usage{
split_at_percentile(x, frac, labels, type = "higher", explicit_na = NA)
}
\arguments{
\item{x}{A vector.}

\item{frac}{A numeric vector with percentiles (between 0 and 1) at which to split
the vector. Alternatively, use character strings "median", "tertile",
"quartile", "quintile", or "decile" for common splits.}

\item{labels}{Vector with factor labels.}

\item{type}{"higher" will split group below fraction and last group equal
above last fraction. and "lower" will split group below fraction (vs. equal
and above).}

\item{explicit_na}{If not NA, NAs will be recoded as a factor level of the
provided name. If TRUE, the name will default to '(Missing)'.}
}
\value{
A vector of type factor with two levels.
}
\description{
Splits a numeric vector into groups at specified percentiles. Values below
(or above) the percentiles are assigned to one group, values equal to or
above (or below) the percentiles are assigned to the other groups. Optionally
missing values can be assigned to a separate factor level.
}
\details{
This function computes the specified percentiles of the input vector and
assigns each value to a group based on these percentiles. The resulting
groups are returned as a factor with the specified labels. The
\code{type} parameter determines whether values equal to the percentile
thresholds are included in the lower or higher group.

Common splits can be specified using character strings for the \code{frac}
parameter:
\itemize{
\item "median": splits at the 50th percentile
\item "tertile": splits at the 33.3rd and 66.6th percentiles
\item "quartile": splits at the 25th, 50th and 75th percentiles
\item "quintile": splits at the 20th, 40th, 60th and 80th percentiles
\item "decile": splits at the 10th, 20th, ..., 90th percentiles
}

The \code{labels} parameter should contain one more label than the number of
percentiles specified in \code{frac}, as it defines the labels for each
resulting group.

If \code{explicit_na} is provided (not NA), missing values in the input
vector will be recoded as a separate factor level with the specified name.
If \code{explicit_na} is set to TRUE, the name will default to '(Missing)'.
}
\examples{
## Generate sample data
x <- sample(c(1:100, NA), 1000, replace = TRUE) 

## Ternary split
split_at_percentile(x, "tertile", explicit_na = TRUE) |> table()

## Quartile split with custom labels
split_at_percentile(
  x,
  frac = c(0.25, 0.5, 0.75), 
  labels = c("0-24", "25-49", "50-74", "75-100")
) |> table()

## Quintile split
split_at_percentile(x, frac = "quintile") |> table() |> prop.table() |> round(2)
 
}
