% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackDF.R
\name{stackDF}
\alias{stackDF}
\title{Stack a data.frame}
\usage{
stackDF(df, index = 1)
}
\arguments{
\item{df}{a data.frame object with at least 3 columns to be stacked.}

\item{index}{a single integer to specify the column that serves as the index for stacking.}
}
\value{

}
\description{
Stack a data.frame object with a given column index.
}
\examples{
\dontrun{
stackDF(df, index = 1)
}
}
