% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_table.R
\name{t_test_table}
\alias{t_test_table}
\title{Creats a t.test tabel for multiple dependent variables}
\usage{
t_test_table(
  dv,
  iv,
  data,
  conditions = levels(factor(iv))[1:2],
  labels = NULL,
  concise = TRUE,
  nice_p = TRUE,
  digits = 1,
  var_equal = FALSE,
  label_attr = TRUE,
  manova = TRUE,
  order = "12",
  type = "df",
  caption = "t-test table",
  bootstrap_options = c("condensed", "striped", "hover"),
  full_width = TRUE
)
}
\arguments{
\item{dv}{A data frame with the dependent variables or a character vector
with variable names when data is defined.}

\item{iv}{A data frame or vector with the independent variable or a character
if data is defined.}

\item{data}{A data frame.}

\item{conditions}{A character vectot of length two with the names of the two
conditions. Defaults to the first two levels of the independent variable
'iv' if applicable.}

\item{labels}{A character vector of length two with labels for the dependent
variables.}

\item{concise}{A more concise table with mean and sd in one column.}

\item{nice_p}{If TRUE, p values are printed in a nice format.}

\item{digits}{Number of digits for rounding mean and sd values}

\item{var_equal}{If FALSE, a t-test for unequal variances is calculated.}

\item{order}{Either "12" or "21" depicting whether group two is compared to
group one or vice versa.}

\item{type}{Either "df" for data frame or "html" for html table.}

\item{caption}{Tabel caption is type = "html"}

\item{bootstrap_options}{see kable_styling()}

\item{full_width}{see kable_styling()}
}
\value{
A tibble or a kableExtra object
}
\description{
Creats a t.test tabel for multiple dependent variables
}
\examples{
dv <- data.frame(
  a = c(runif(85) * 150, runif(115) * 100),
  b = runif(200) * 100
  )
iv <- factor(c(rep("A", 85), rep("B", 115)))
t_test_table(dv, iv, labels = c("Motivation", "Achievement"))
}
