% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstackDF.R
\name{unstackDF}
\alias{unstackDF}
\title{Unstacking a data.frame object}
\usage{
unstackDF(df, row_id, col_id, value_id, na_value = NA)
}
\arguments{
\item{df}{The input data.frame object}

\item{row_id}{An integer indicating the column id which specifies the rows of the unstacked matrix}

\item{col_id}{An integer indicating the column id which specifies the columns of the unstacked matrix}

\item{value_id}{An integer indicating the column id which specifies the values to be filled in the unstacked matrix.}

\item{na_value}{Specifies whether to replace NA values, such as from missing row/column combinations. Defaults to "NA" (no replacement).}
}
\value{

}
\description{
This function reliably unstacks a data frame into a matrix
}
