% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_self.R
\name{update_self}
\alias{update_self}
\title{Update this package from GitHub if a newer version is available}
\usage{
update_self()
}
\description{
Update this package from GitHub if a newer version is available
}
