% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-to-colors.R
\name{values_to_colors_discrete}
\alias{values_to_colors_discrete}
\title{Map discrete values to colors.}
\usage{
values_to_colors_discrete(x, values, colors)
}
\arguments{
\item{x}{Numeric or factor vector. Can be named or unnamed, but must align
with \code{values} by name (preferred) or by position (same length).}

\item{values}{Numeric or factor vector of values to map (same type as \code{x}).}

\item{colors}{Character vector of hex color codes (same length as \code{values}
and in the same order).}
}
\value{
Named character vector of hex colors with names equal to \code{x}.
}
\description{
Map discrete values to colors.
}
\examples{
 x <- c("A", "B", "C", "A", "B")
 values <- c("A", "B", "C")
 colors <- c("#7f0000", "#ffffff", "#006400")
 result <- values_to_colors_discrete(x, values, colors)
 print(result)
}
