\name{wmisc-package}
\alias{wmisc-package}
\alias{wmisc}
\docType{package}
\title{
\packageTitle{wmisc}
}
\description{
\packageDescription{wmisc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{wmisc}
\packageIndices{wmisc}
}
\author{\packageAuthor{wmisc}
Maintainer: \packageMaintainer{wmisc}
}
\keyword{ package }
