% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtab.R
\name{wtab}
\alias{wtab}
\title{Weighted tabulation}
\usage{
wtab(dat, row, col, by, wt, ...)
}
\arguments{
\item{dat}{A data frame}

\item{row}{Row variable}

\item{col}{Column variable}

\item{by}{Grouping (super row) variable}

\item{wt}{Weight variable}
}
\description{
`wtab` produces one-way, two-way and three-way tabulations with weighted data.
}
\details{
This is actually a wrapper function that calls other ones. `wtab` is compatible
with the `%>%` pipe operator
}
