% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtab.R
\name{wtab_1way}
\alias{wtab_1way}
\title{One-way Weighted tabulation}
\usage{
wtab_1way(dat, row, wt)
}
\arguments{
\item{dat}{A data frame}

\item{row}{Row variable}

\item{wt}{Weight variable}
}
\description{
wtab_1way produces one-way with weighted data.
}
