% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_names.R
\name{x_names}
\alias{x_names}
\alias{x_nanmes}
\title{x_names}
\usage{
x_names(list, width = 70)
}
\arguments{
\item{list}{data-frame or another object of class list}

\item{width}{Linewidth before linebreak}
}
\value{
A character string with variable names formated as "c(naem, name, name, ...)"
}
\description{
Give a list of variable names to the CLIPBOARD and the screen
}
