% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreement_analysis.R
\name{agreement_analysis}
\alias{agreement_analysis}
\alias{nice_agreement_table}
\title{Perform within agreement analysis}
\usage{
agreement_analysis(
  data,
  vars = names(data),
  grouping,
  rv,
  crit = 0.7,
  min_group_size = 10,
  auto_labels = TRUE,
  type = "df",
  labels = NULL,
  n_sim = 10000
)

nice_agreement_table(...)
}
\arguments{
\item{data}{A data-frame}

\item{vars}{Vector of character strings with names of variables to be
analyzed}

\item{grouping}{Grouping variable. Either as a character string or a vector}

\item{rv}{Number of response options for calculating random variance}

\item{crit}{Critical agreement level.}

\item{min_group_size}{Minimal group size. Smaller groups are excluded.}

\item{auto_labels}{If TRUE, variable names are taken from a label attribute.}

\item{labels}{Character vector with new variable labels.}

\item{n_sim}{Number of simulations for the agreement analyses}
}
\value{
A Data-frame
}
\description{
This function performs an agreement analysis including intraclass correlation
coefficients, group-related variance, and related diagnostics for items with
a nominal or ordinal scale. The function outputs a data.frame or a HTML table
giving the results.
}
\examples{

vars <- c("Norm_Pop_academic", "Norm_Pop_conform", "Norm_Pop_sporty", 
  "Norm_Pop_cool_clothes", "Norm_Pop_helpful", "Norm_Unpop_academic", 
  "Norm_Unpop_nonconform", "Norm_Unpop_sporty", "Norm_Unpop_cool_clothes", 
  "Norm_Unpop_language", "Norm_LM_academic", "Norm_LM_conform", 
  "Norm_LM_sporty", "Norm_LM_cool_clothes", "Norm_LM_helpful", 
  "Norm_LL_academic", "Norm_LL_nonconform", "Norm_LL_sporty", "Norm_LL_cool_clothes", 
  "Norm_LL_language")
  
nice_agreement_table(
    data = wmisc:::ex_agreement, 
    vars = vars, 
    grouping = "id_class_teacher", 
    rv = 4, 
    crit = 0.6, 
    n_sim = 100)
}
