% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_CI.R
\name{alpha_CI}
\alias{alpha_CI}
\title{Cronbachs Alpha Confidence Interval}
\usage{
alpha_CI(alpha, n, items, ci)
}
\arguments{
\item{alpha}{Alpha value}

\item{n}{Sample size}

\item{items}{Number of items}

\item{ci}{Confidence level (e.g. 0.95 for 95 percent)}
}
\description{
Cronbachs Alpha Confidence Interval
}
