% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.R
\name{average}
\alias{average}
\title{Average function with minimal valid values or maximum NAs}
\usage{
average(x, min_valid, max_na, func = mean)
}
\arguments{
\item{x}{Vector of numeric values.}

\item{min_valid}{Minimal number of valid values that is required for
calculating the mean. A value between 0 and 1 indicates a proportion of
values (e.g., 0.5 = 50 percent of values have to be valid).}

\item{max_na}{Maximum number of NAs that are allowed before returning NA. A
value between 0 and 1 indicates a proportion of values (e.g., 0.5 = 50
percent NAs are allowed).}

\item{func}{A function for calculating the average. Default is \code{mean()}.}
}
\value{
A numeric value representing the average, or NA if the conditions are not met.
}
\description{
Computes arithmetic means (or other averages) with minimal requirements
for valid values or maximum allowed NAs.
}
\details{
If the number of valid values is below \code{min_valid} or the number of NAs is
above \code{max_na}, the function returns NA. Otherwise, it computes the average using
the specified function \code{func}.

If both \code{min_valid} and \code{max_na} are
provided, both conditions have to be met for the calculation to proceed.
Missing values (NAs) are ignored in the calculation of the average.
If the conditions for \code{min_valid} or \code{max_na} are not met,
NA is returned.
}
\examples{
dat <- c(1:5, NA, NA)
average(dat, min_valid = 5)
average(dat, min_valid = 6)
average(dat, max_na = 2)
average(dat, max_na = 1)
average(dat, min_valid = 0.5, func = median)
average(dat, min_valid = 0.9)
}
\author{
Juergen Wilbert
}
