% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_values.R
\name{change_values}
\alias{change_values}
\title{Change Values}
\usage{
change_values(x, ..., .default = NULL)
}
\arguments{
\item{x}{A vector in which values will be replaced.}

\item{...}{An ellipsis of pairs of values to be replaced. Each pair consists
of the original value and the corresponding replacement value.}

\item{.default}{The default value to be used for elements not specified in
the replacement pairs. If not provided, elements not specified will remain
unchanged.}
}
\value{
A modified vector with replaced values.
}
\description{
Replace specific values in a vector with new values.
}
\examples{
## Retains NAs
change_values(c(1, 2, 3, NA), 2 ~ "two", 3 ~ "three")

## Set a specific default value
change_values(c(1, 2, 3), 2 ~ "two", .default = "default")

## No error message when no value to convert is found
change_values(c(1, 2, 3), 4 ~ "four")

## Recode NAs and implicit conversion from numeric to character
change_values(c(NA, 1, NA, 2), NA ~ "This is a missing value")

}
