% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_values.R
\name{change_values}
\alias{change_values}
\title{Change Values}
\usage{
change_values(x, ..., .default = NULL)
}
\arguments{
\item{x}{A vector in which values will be replaced.}

\item{...}{An ellipsis of pairs of values to be replaced. Each pair consists of the original value and the corresponding replacement value.}

\item{.default}{The default value to be used for elements not specified in the replacement pairs. If not provided, elements not specified will remain unchanged.}
}
\value{
A modified vector with replaced values.
}
\description{
Replace specific values in a vector with new values.
}
\examples{
change_values(c(1, 2, 3), 2 ~ "two", 3 ~ "three")
# Output: c(1, "two", "three")

change_values(c(1, 2, 3), 2 ~ "two", .default = "default")
# Output: c("default", "two", "default")

change_values(c(1, 2, 3), 4 ~ "four")
# Output: c("1", "2", "3")

}
