% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attribute.R
\name{get_wmisc_attributes}
\alias{get_wmisc_attributes}
\alias{set_wmisc_attributes}
\title{Access or Modify "wmisc" Attributes}
\usage{
get_wmisc_attributes(x)

set_wmisc_attributes(x, ...)
}
\arguments{
\item{x}{An R object (usually a list) from or to which the \code{"wmisc"} attribute is accessed or modified.}

\item{...}{Named values to add or update in the \code{"wmisc"} attribute.}
}
\value{
\code{get_wmisc_attributes()} returns the \code{"wmisc"} attribute of \code{x}, or \code{NULL} if none is present.
\code{set_wmisc_attributes()} returns \code{x} with the updated \code{"wmisc"} attribute.
}
\description{
These functions access or modify the \code{"wmisc"} attribute of an R object.
The attribute is typically used to store metadata in \code{wmisc}-compatible structures.
}
\details{
\itemize{
\item \code{get_wmisc_attributes()} extracts and returns the \code{"wmisc"} attribute. Returns \code{NULL} if it does not exist.
\item \code{set_wmisc_attributes()} adds or updates entries in the \code{"wmisc"} attribute. If a name occurs multiple times, the last value is retained.
}
}
\examples{
obj <- list(a = 1, b = 2)
obj <- set_wmisc_attributes(obj, key1 = "value1", key2 = "value2")
get_wmisc_attributes(obj)
obj <- set_wmisc_attributes(obj, key1 = "new_value1", key3 = "value3")
get_wmisc_attributes(obj)

}
\seealso{
\code{\link[=attributes]{attributes()}}, \code{\link[=attr]{attr()}}
}
