% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_corr.R
\name{mean_corr}
\alias{mean_corr}
\title{Average correlations by Fisher Z transformation}
\usage{
mean_corr(r)
}
\arguments{
\item{r}{A vector of correlation coeficients}
}
\value{
The average correlation
}
\description{
Average correlations by Fisher Z transformation
}
