% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_R2.R
\name{multilevel_r2}
\alias{multilevel_r2}
\title{Multilevel R squared for two-level models}
\usage{
multilevel_r2(model1, model2, n = table(model1$groups))
}
\arguments{
\item{model1}{The first model.}

\item{model2}{The second model.}

\item{n}{Optional vector of group sizes for level 2.}
}
\value{
A list with the following components: \item{r2_l1}{The R squared
value for level 1.} \item{r2_l2}{The R squared value for level 2.}
\item{r2_l1_2}{The combined R squared value for levels 1 and 2.}
\item{r2_l2_2}{The adjusted R squared value for levels 1 and 2 (with
optional group sizes).} \item{n}{The harmonic mean of the group sizes (if
provided).}
}
\description{
Calculates the multilevel R squared for two-level models.
}
\examples{
data(sleepstudy, package = "lme4")
model1 <- nlme::lme(Reaction ~ 1, data = sleepstudy, random =~ 1|Subject)
model2 <- nlme::lme(Reaction ~ Days, data = sleepstudy, random =~ 1|Subject)
multilevel_r2(model1, model2, table(sleepstudy$Subject))
}
\references{
Snijders, T. A. B., & Bosker, R. J. (1994). Modeled Variance in
Two-Level Models. Sociological Methods & Research, 22(3), 342-363.
}
