% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_corrmatrix.R
\name{nice_corrmatrix}
\alias{nice_corrmatrix}
\title{Nice correlationmatrix}
\usage{
nice_corrmatrix(cr, upper = TRUE, lower = TRUE, alpha = 0.1,
  digits = 2, n_sig, values = TRUE, stars = TRUE, autocor = "-",
  square = FALSE, caption = "Correlation matrix.",
  sig.10 = "<U+271D>  ", type = "df")
}
\arguments{
\item{cr}{A correlation matrix returned by the psych package}

\item{upper}{TRUE if upper triangle should be included}

\item{lower}{TRUE if lower triangle should be included}

\item{alpha}{Alpha level}

\item{digits}{Round to given digit position}

\item{n_sig}{Charcter indexing non-significant values}

\item{values}{TRUE if values should be included}

\item{stars}{TRUE if stars should be included}

\item{autocor}{Character for diagonal (e.g. "-" or "1.00")}

\item{square}{Report squared r instead of r}

\item{caption}{Caption for a html table.}

\item{sig.10}{Character indexing .10 sognificance level}

\item{type}{Charcater string. "df" for data-frame. "html" for html table (needs knitr and kableExtra packages)}
}
\value{
A data-frame or a html table object
}
\description{
From a psych corr object it gives a correlationmatrix formated for overview
}
