% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_descriptives.R
\name{nice_descriptives}
\alias{nice_descriptives}
\title{Table with descriptive statistics}
\usage{
nice_descriptives(
  data,
  auto_labels = FALSE,
  title = "Descriptive statistics",
  footnote = NULL,
  file = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame}

\item{auto_labels}{If TRUE, variable names are taken from a label attribute.}

\item{title}{Title for the table.}

\item{footnote}{Footnote for the table.}

\item{file}{If a file name is provided, the table is saved to this file.}

\item{...}{Further arguments passed to \code{\link[=nice_table]{nice_table()}}.}
}
\value{
A data frame with descriptive statistics
}
\description{
Table with descriptive statistics
}
\examples{
nice_descriptives(mtcars)

nice_descriptives(wmisc:::mtcars_labeled, auto_labels = TRUE, round = 2)
}
