% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_descriptives.R
\name{nice_descriptives}
\alias{nice_descriptives}
\title{Table with descriptive statistics}
\usage{
nice_descriptives(
  data,
  round = NULL,
  auto_labels = FALSE,
  title = "Descriptive statistics",
  footnote = NULL,
  file = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame}

\item{round}{Digits for round function}

\item{auto_labels}{If TRUE, variable names are taken from a label attribute.}
}
\value{
A data frame with descriptive statistics
}
\description{
Table with descriptive statistics
}
\examples{
nice_descriptives(mtcars)

nice_descriptives(wmisc:::mtcars_labeled, auto_labels = TRUE, round = 2)
}
