% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_descriptives.R
\name{nice_descriptives}
\alias{nice_descriptives}
\title{Table with descriptive statistics}
\usage{
nice_descriptives(
  data,
  round = NULL,
  labels = FALSE,
  title = "Descriptive statistics",
  note = "MAD is the Median average deviation"
)
}
\arguments{
\item{data}{A data frame}

\item{round}{Digits for round function}

\item{labels}{Item labels.}
}
\value{
A data frame with descriptive statistics
}
\description{
Table with descriptive statistics
}
\examples{
nice_descriptives(mtcars)
}
