% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_frequencies.R
\name{nice_frequencies}
\alias{nice_frequencies}
\title{Table with frequency statistics}
\usage{
nice_frequencies(
  data,
  grouping = NULL,
  label = NULL,
  label_grouping = NULL,
  show_missing = TRUE,
  show_percent = TRUE,
  percent_base = "column",
  add_total = TRUE,
  auto_labels = TRUE,
  title = NULL,
  footnote = NULL,
  file = NULL,
  round = 1,
  ...
)
}
\arguments{
\item{data}{Vector.}

\item{grouping}{A grouping variable as a vector.}

\item{label}{Set label for the variable name.}

\item{label_grouping}{Set label for the grouping variable name.}

\item{show_missing}{If TRUE, adds a row for the number of missing values.}

\item{show_percent}{If TRUE, adds a column for percentages.}

\item{percent_base}{If show_percent is TRUE, this argument specifies the base for percentage calculations.
Options are "column" (percentages calculated within each column), "row"
(percentages calculated within each row), or "total" (percentages calculated based on the total count).
Default is "column".}

\item{auto_labels}{If TRUE, variable names are taken from a label attribute.}

\item{title}{Table title.}

\item{footnote}{Table footnote.}

\item{file}{Filename.}

\item{...}{Further arguments passed to \code{\link[=nice_table]{nice_table()}}.}
}
\value{
An html table with frequencies
}
\description{
Table with frequency statistics
}
\examples{
nice_frequencies(wmisc:::mtcars_labeled[[11]])
nice_frequencies(
  wmisc:::mtcars_labeled$cyl, 
  wmisc:::mtcars_labeled$am
)
nice_frequencies(
 wmisc:::mtcars_labeled$cyl,
 wmisc:::mtcars_labeled$am,
 show_missing = FALSE,
 show_percent = FALSE
)
nice_frequencies(
  wmisc:::mtcars_labeled$cyl,
  wmisc:::mtcars_labeled$am,
  label = "Cylinders",
  label_grouping = "Transmission",
  title = "Cylinders by Transmission Type"
)
}
