% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_frequencies.R
\name{nice_frequencies}
\alias{nice_frequencies}
\title{Table with frequency statistics}
\usage{
nice_frequencies(
  data,
  label = NULL,
  show_missing = TRUE,
  auto_labels = TRUE,
  title = NULL,
  footnote = NULL,
  file = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame}

\item{label}{Set label for the variable name.}

\item{show_missing}{If TRUE, adds a row for the number of missing values.}

\item{auto_labels}{If TRUE, variable names are taken from a label attribute.}

\item{title}{Table title.}

\item{footnote}{Table footnote.}

\item{file}{Filename.}

\item{...}{Further arguments passed to \code{\link[=nice_table]{nice_table()}}.}
}
\value{
An html table with frequencies
}
\description{
Table with frequency statistics
}
\examples{
nice_frequencies(mtcars[[11]])
}
