% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_item_analysis.R
\name{nice_item_analysis}
\alias{nice_item_analysis}
\title{Table with detailed item statistics for one scale}
\usage{
nice_item_analysis(
  data,
  scales,
  labels = NULL,
  round = 2,
  ci = TRUE,
  conf_level = 0.95,
  check_key = TRUE,
  keys = NULL,
  keys_from_weights = FALSE,
  RMSEA = FALSE,
  difficulty = FALSE,
  values = NULL,
  fa = TRUE,
  ...
)
}
\arguments{
\item{data}{A data Frame}

\item{round}{Rounds values to given decimal position.}

\item{ci}{If TRUE confidence intervals are calculated.}

\item{conf_level}{Confidence level (e.g. 0.95 for 95 percent).}

\item{check_key}{Check_key for the psych::alpha function.}

\item{keys}{Optional key argument for the psych::alpha function.}

\item{keys_from_weights}{If TRUE, tries to define keys from the weights
dictionary attribute. These are only available when you defined them with
the scaledic package.}

\item{RMSEA}{If TRUE RMSEA is calculated.}

\item{difficulty}{If TRUE, the difficulty of the item is calculated.}

\item{values}{Sets maximum and minimum valid values necessary to calculate
item difficulty.}

\item{fa}{If TRUE, a one factor exploratory factor analyses is calculated and
loadings are reported.}

\item{scale}{A vector with variable names that define the scale.}
}
\value{
A data frame with concise scale indices.
}
\description{
Returns a data.frame with detailed item analyses for the provided scale.
}
\examples{
nice_item_analysis(
  wmisc:::data_emo, 
  scale = wmisc:::data_emo_scales[[1]], 
  difficulty = TRUE, 
  values = c(1,5)
)
}
