% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_loadings.R
\name{nice_loadings}
\alias{nice_loadings}
\alias{nice_efa}
\title{Creates table of loadings of exploratory factor analysis based on the
psych::fa function}
\usage{
nice_loadings(
  x,
  factor_names = NULL,
  sort = TRUE,
  cut = 0.2,
  round = 2,
  title = "Loading matrix",
  footnote = NULL,
  auto_labels = TRUE
)

nice_efa(..., file = NULL)
}
\arguments{
\item{x}{Object returned from the psych::fa function.}

\item{factor_names}{A character vector with names for the resulting factors.
If not provided, default names are chosen.}

\item{sort}{If TRUE, loadings are sorted.}

\item{cut}{Loadings below cut will be omitted in the resulting data frame.}

\item{round}{Number of digits to round loadings (based on the base::round
function)}
}
\value{
A data.frame
}
\description{
Creates table of loadings of exploratory factor analysis based on the
psych::fa function
}
\examples{
mtcars |> psych::fa(nfactors = 2) |> nice_efa()

}
