% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_p.R
\name{nice_p}
\alias{nice_p}
\title{Format p values}
\usage{
nice_p(p, equal_sign = FALSE, digits = 3)
}
\arguments{
\item{p}{Vector of p values.}

\item{equal_sign}{If TRUE an equal sign is set before p values.}
}
\value{
A character vector with nicely formatted p values
}
\description{
Format p values
}
\examples{
p <- c(0.04, 0.9, 0.10, 0.001, 1, NA)
nice_p(p)
paste0("p", nice_p(p, equal_sign = TRUE))
}
