% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_regression.R
\name{nice_regression_table}
\alias{nice_regression_table}
\title{Create a nice table from one or more regression models}
\usage{
nice_regression_table(
  ...,
  digits = 2,
  labels_models = NULL,
  rename_labels = list(),
  rename_cols = list(),
  auto_col_names = TRUE
)
}
\description{
Create a nice table from one or more regression models
}
\examples{
lm(mpg ~ am + disp + hp, data = mtcars) |> 
  nice_regression_table()
  
nice_regression_table(
  nlme::lme(mpg~disp, data = mtcars, random = ~1|am),
  nlme::lme(mpg~disp + hp, data = mtcars, random = ~1|am)
)   
  
nice_regression_table(
  wmisc:::model_lmer_1, wmisc:::model_lmer_2,
  rename_labels = list(
    "EffectTrend" = "Trend", "EffectSlope" = "Slope", "TimePost" = "Post", 
    "ConditionTraining" = "Training", "id_subject" = "Subject"),
  rename_cols = list("Estimate" = "B", "SE" = "se"),
  labels_models = c("Only pretest", "Pre- and posttest") 
)
}
