% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_regression.R
\name{nice_regression_table}
\alias{nice_regression_table}
\title{Create a nice table from one or more regression models}
\usage{
nice_regression_table(
  ...,
  digits = 2,
  labels_models = NULL,
  rename_labels = list(),
  rename_cols = list(),
  auto_col_names = TRUE
)
}
\description{
Create a nice table from one or more regression models
}
\examples{
lm(mpg ~ am + disp + hp, data = mtcars) |> 
  nice_regression_table()
nice_regression_table(
  wmisc:::model1, wmisc:::model2,
  rename_labels = list(
    "EffectTrend" = "Trend", "EffectSlope" = "Slope", "TimePost" = "Post", 
    "ConditionTraining" = "Training"),
  rename_cols = list("Estimate" = "b", "SE" = "se"),
  labels_models = c("Only pretest", "Pre- and posttest") 
)
}
