% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_sem.R
\name{nice_sem}
\alias{nice_sem}
\title{Nice html table for lavaan structure equation model fitted objects}
\usage{
nice_sem(
  x,
  standardized = TRUE,
  fitmeasures = c(CFI = "cfi", TLI = "tli", RMSEA = "rmsea", `RMSEA ci lower` =
    "rmsea.ci.lower", AIC = "aic", BIC = "bic"),
  remove_cols = NULL,
  round = 2,
  title = "Structure equation model",
  footnote = NULL,
  file = NULL
)
}
\arguments{
\item{x}{An object returned from the lavaan sem or cfa function.}

\item{remove_cols}{Either column number or column names to be removed.}

\item{title}{Title string.}

\item{footnote/note}{Add footnote}
}
\description{
Nice html table for lavaan structure equation model fitted objects
}
\examples{
nice_sem(wmisc::lavaan_fit)
}
