% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_table.R
\name{nice_table.default}
\alias{nice_table.default}
\alias{nice_table}
\title{Create a nicely formatted table}
\usage{
\method{nice_table}{default}(
  x,
  title = NULL,
  footnote = NULL,
  spanner = NULL,
  row_group = NULL,
  row_group_order = NULL,
  rownames = NULL,
  file = NULL,
  cols_label = NULL,
  cols_align = NULL,
  decimals = NULL,
  round = NULL,
  label_na = NULL,
  markdown = FALSE,
  gt = NULL,
  sort = NULL,
  sort_decreasing = FALSE,
  ...
)
}
\arguments{
\item{x}{The data frame to be formatted into a table}

\item{title}{Title string.}

\item{footnote}{Add footnote}

\item{spanner}{List with information on grouping columns. E.g. \code{spanner = list("M" = 2:3, "SD" = 4:6)}.}

\item{row_group}{List with information on grouping rows \code{row_group = list("Start" = 1:2, "That is the second" = 3:5)}}

\item{row_group_order}{List with information on grouping order.}

\item{rownames}{Logical or \code{NULL}. If TRUE, rownames are shown. If \code{NULL},
rownames are shown when they are not identical to
\code{as.character(1:nrow(x))}.}

\item{file}{Character string with filename. If set, an additional file is
exported (html or docx format is possible). If set \code{TRUE}, a filename is
automatically created based on the title.}

\item{cols_label}{List with renaming information for columns (old_name =
new_name).}

\item{cols_align}{List with align align information. E.g., \code{list(left = c(2,3), right = 1)}.}

\item{decimals}{Number of decimals that will be printed.}

\item{round}{Number of digits to which numbers should be rounded.}

\item{label_na}{= Label for replacing missing values.}

\item{markdown}{If TRUE, interprets cell content as markdown.}

\item{gt}{Additional arguments passed to \code{gt::gt()}}

\item{...}{Various arguments for backward compatibility.}
}
\value{
A gt table object.
}
\description{
This function takes a data frame and formats it into a nicely formatted HTML
table using the \code{gt} packages.
}
\examples{
df <- data.frame(
  x = 1:5, y = rnorm(5, mean = 10, sd = 12),
  c = letters[5:1], d = sample(letters, 5)
)
nice_table(
  df,
  title = "A nice title",
  footnote = c("Footnote 1", "Footnote 2"),
  spanner = list("One" = 1:2, "Two" = 3:4),
  row_group = list("Start" = 1:2, "That is the second" = 3:5),
  cols_label = list(x = "First", y = "Second", c = "Third", d = "Fourth"),
  decimals = 1
)

}
