% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_table.R
\name{nice_table}
\alias{nice_table}
\title{Create a nicely formatted table}
\usage{
nice_table(
  x,
  title = NULL,
  footnote = NULL,
  spanner = header,
  row_group = NULL,
  header = NULL,
  pack = NULL,
  rownames = FALSE,
  file = NULL,
  cols_label = NULL,
  use_labels = TRUE,
  decimals = NULL,
  round = NULL,
  engine = getOption("wmisc.nice.table.engine"),
  extra = NULL,
  gt = NULL,
  kable = NULL
)
}
\arguments{
\item{x}{The data frame to be formatted into a table}

\item{title}{Title string.}

\item{footnote}{Add footnote}

\item{file}{(works only when `engine = "gt"`) If set, an additional file with
the table is produced.}

\item{cols_label}{List with renaming information for columns (old_name =
new_name).}

\item{decimals}{Number of decimals that will be printed.}

\item{round}{Number of digits to which numbers should be rounded.}

\item{extra}{Additional arguments passed to `kableExtra::kable_classic_2()`}

\item{gt}{Additional arguments passed to `gt::gt()`}

\item{kable}{Additional arguments passed to `knitr::kable()`}
}
\value{
A nicely formatted HTML table
}
\description{
This function takes a data frame and formats it into a nicely formatted HTML
table using the `knitr` and `kableExtra` packages.
}
\examples{
df <- data.frame(x = 1:5, y = rnorm(5, mean = 10, sd = 12), c = letters[5:1], d = sample(letters, 5))
nice_table(
  df,
  title = "A nice title",
  footnote = c("Footnote 1", "Footnote 2"),
  spanner = list("One" = 1:2, "Two" = 3:4),
  row_group = list("Start" = 1:2, "That is the second" = 3:5),
  cols_label = list(x = "First", y = "Second", c = "Third", d = "Fourth"),
  decimals = 1
)

}
