% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_table.R
\name{nice_table}
\alias{nice_table}
\title{Create a nicely formatted table}
\usage{
nice_table(x, ..., extra = NULL)
}
\arguments{
\item{x}{The data frame to be formatted into a table}

\item{...}{Additional arguments passed to `knitr::kable()`}

\item{extra}{Additional arguments passed to `kableExtra::kable_classic_2()`}
}
\value{
A nicely formatted HTML table
}
\description{
This function takes a data frame and formats it into a nicely
formatted HTML table using the `knitr` and `kableExtra` packages.
}
\examples{
df <- data.frame(x = 1:5, y = rnorm(5))
nice_table(df, extra = list(full_width = FALSE))

}
