% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent}
\alias{percent}
\title{Percent of TRUE values}
\usage{
percent(x, round = 2, min_valid, max_na)
}
\arguments{
\item{x}{A vector}

\item{round}{Demical positions (default = 2).}

\item{min_valid}{Minimal number of valid values that is requiered for calculation.
A value between 0 and 1 indicates a proportion of values (e.g., 0.5 = 50 percent of values have to be valid).}

\item{max_na}{Maximum number of NAs that are allowed before returning NA.
A value between 0 and 1 dicates a proportion of values (e.g., 0.5 = 50 percent NAs are allowed).}
}
\value{
A percentage
}
\description{
Returns percentage of the occurance of TRUE values in a vector of logical values.
}
\examples{
percent(runif(10000) > 0.50)
}
