% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values-to-colors.R
\name{plot_continuous_legend}
\alias{plot_continuous_legend}
\title{Plot a continuous color legend for the diverging color scheme.}
\usage{
plot_continuous_legend(
  min = -3,
  max = 3,
  neutral_min = -0.5,
  neutral_max = 0.5,
  low = "#7f0000",
  mid = "#ffffff",
  high = "#006400",
  n = 801,
  unit = 1,
  title = "Z"
)
}
\arguments{
\item{min}{Minimum value that maps to the darkest red (values below are
clipped).}

\item{max}{Maximum value that maps to the darkest green (values above are
clipped).}

\item{neutral_min}{Minimum value of the white "dead zone" (values above are
colored white).}

\item{neutral_max}{Maximum value of the white "dead zone" (values below are
colored white).}

\item{low}{Color for low end (default dark red).}

\item{mid}{Color for midpoint (default white).}

\item{high}{Color for high end (default dark green).}

\item{n}{Number of discrete colors used internally for the gradient.}

\item{title}{Title for the legend.}
}
\value{
A plot of the color legend.
}
\description{
Plot a continuous color legend for the diverging color scheme.
}
