% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotalphabeta.R
\name{plotalphabeta}
\alias{plotalphabeta}
\title{Plot Alpha / Beta graph
Generates a graphic for visualizing alpha and beta error}
\usage{
plotalphabeta(m = 0, sd = 1, d = 2, alpha = 0.05, fill = TRUE,
  col_fill = c("grey75", "grey50"))
}
\arguments{
\item{m}{Mean}

\item{sd}{Standard Deviation}

\item{d}{Effect size}

\item{alpha}{Alpha Level}

\item{fill}{If TRUE, intersections are filled out in grey}

\item{col_fill}{Vector of length two with colornames to fill area below curve}
}
\description{
Plot Alpha / Beta graph
Generates a graphic for visualizing alpha and beta error
}
\examples{
plotalphabeta()

}
