% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_by_list.R
\name{rename_by_list}
\alias{rename_by_list}
\title{Rename variables based on a list of new and old variable names}
\usage{
rename_by_list(data, file = NULL, to_from = NULL, to = NULL, from = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{file}{A filename of an Excel file or a data.frame containing at least
two columns for renaming (variables to rename and target names).}

\item{to_from}{When a filename or a data.frame is provided, a named character
with the names of the target  and source variable names (e.g., c("to" =
"from")). When no filename is provided, to_from must be a vector with named
variable names c("to1" = "from1", "to2" = "from2")).}

\item{to}{When a filename or a data.frame is provided, the name of the column
with the target variable names. When no filename is provided, to must be a
vector with target variable names.}

\item{from}{When a filename or a data.frame is provided, the name of the
column with the source variable names. When no filename is provided, from
must a vector with source variable names.}
}
\value{
A data frame with renamed variables.
}
\description{
Rename the variable in a dataset based of a provided renaming list. The
renaming list can be provided as an Excel file or as a data.frame with at
least two columns (one with the source variable names and one with the target
variable names). Alternatively, the renaming list can be provided directly as
named vectors.
}
\details{
This function is useful when you want to rename multiple variables
in a data frame based on a predefined list of new and old variable names.
You can provide the renaming list as an Excel file or as a data.frame,
making it easy to manage and update the renaming list externally. The
function reads the renaming list, extracts the relevant columns, and applies
the renaming to the data frame. It also provides feedback on the number of
variables that were renamed.

If both \code{file} and \code{to_from} are provided, the function will read the
renaming list from the specified file and use the columns specified in
\code{to_from} to determine the source and target variable names. If only \code{file}
is provided, the function will use the columns specified in \code{to} and \code{from}
to determine the source and target variable names.
If \code{file} is NULL and \code{to_from} is provided, the function will use the named
vector in \code{to_from} to determine the source and target variable names.
If no variables in the renaming list match the variable names in the data
frame, the function will stop and return an error message indicating that
no variables were renamed.
}
\examples{
dat <- data.frame(A = NA, B = NA, C = NA, D = NA)
rename_by_list(dat, to = c("albert", "bea"), from = c("A", "B"))
rename_by_list(dat, to_from = c("carl" = "C", "daniel" = "D"))
dic <- data.frame(old = c("A", "B"), new = c("albert", "bea"))
rename_by_list(dat, dic, to_from = c("new" = "old"))
rename_by_list(dat, dic, to = "new", from = "old")
\dontrun{
rename_by_list(dat, "rename_list,xlsx", to_from = c("new" = "old"))
rename_by_list(dat, "rename_list,xlsx", to = "new", from = "old")
}
}
\author{
Juergen Wilbert
}
\keyword{internal}
