% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attribute.R
\name{rename_from_labels}
\alias{rename_from_labels}
\title{Renames the variables from a data frame to a label that is provided in the label attribute}
\usage{
rename_from_labels(data, keep = FALSE)
}
\arguments{
\item{data}{A dataframe}
}
\value{
A dataframe wit renamed labels
}
\description{
Renames the variables from a data frame to a label that is provided in the label attribute
}
