% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_numeric.R
\name{round_numeric}
\alias{round_numeric}
\title{This function rounds numeric columns in a data frame to a specified number of digits.}
\usage{
round_numeric(x, digits = 0)
}
\arguments{
\item{x}{A data frame.}

\item{digits}{Number of digits to round to. Defaults to 0.}
}
\value{
A modified data frame with rounded numeric columns.
}
\description{
This function rounds numeric columns in a data frame to a specified number of digits.
}
\examples{
# Create a sample data frame
data <- data.frame(
  numeric_col1 = c(1.234, 2.345, 3.456),
  numeric_col2 = c(4.567, 5.678, 6.789),
  non_numeric_col = c("A", "B", "C")
)
attr(data, "my") <- "It keeps all the attributes!"
# Round numeric columns in the data frame
round_numeric(data, digits = 2)
attr(data, "my") 
}
\seealso{
\code{\link{round}}
}
