% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private-row_df.R
\name{row_df}
\alias{row_df}
\title{Read a single row of data into a data frame}
\usage{
row_df(x, sep = "|")
}
\arguments{
\item{x}{A character string representing a row of data}

\item{sep}{The separator used in the data string. Default is "|"}
}
\value{
A data frame with one row
}
\description{
This function takes a character string representing a single row of data,
with columns separated by a specified separator, and converts it into a data
frame.
}
\examples{
row_df("id|name|age\n1|Alice|30")
row_df("id,name,age\n3,Charlie,35", sep = ",")
}
\keyword{internal}
