% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private-row_df.R
\name{row_df}
\alias{row_df}
\title{Read a single row of data into a data frame}
\usage{
row_df(x, sep = "|")
}
\arguments{
\item{x}{A character string representing a row of data}

\item{sep}{The separator used in the data string. Default is "|"}
}
\description{
Read a single row of data into a data frame
}
