% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_at_percentile.R
\name{split_at_percentile}
\alias{split_at_percentile}
\title{Factorizes a vector by percentiles}
\usage{
split_at_percentile(
  x,
  frac = c(1/3, 2/3),
  labels = c("low", "middle", "high"),
  type = "higher",
  explicit_na = NA
)
}
\arguments{
\item{x}{A vector}

\item{frac}{Fractions to split at (e.g., `c(1/3, 2/3)` will result in three
groups and 0.5 will result in a median split)}

\item{labels}{Vector with factor labels.}

\item{type}{"higher" will split group below fraction and last group equal
above last fraction. and "lower" will split group below fraction (vs. equal
and above)}

\item{explicit_na}{If not NA, NAs will be recoded as a factor level of the
provided name. If TRUE, the name will default to '(Missing)'.}
}
\value{
A vector of type factor with two levels.
}
\description{
Factorizes a vector by percentiles
}
\examples{
x <- sample(c(1:10, NA), 100, replace = TRUE)
x <- split_at_percentile(x, explicit_na = TRUE)
table(x)

split_at_percentile(
  1:100, 
  frac = c(0.25, 0.5, 0.75), 
  labels = c("0-24", "25-49", "50-74", "75-100")
)

}
