% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_table.R
\name{t_test_table}
\alias{t_test_table}
\title{Creats t.test tabel for multiple dependent variables}
\usage{
t_test_table(dv, iv, conditions = levels(factor(iv))[1:2],
  labels = NULL, concise = TRUE, nice_p = TRUE, digits = 1,
  var_equal = FALSE, order = "12", type = "df")
}
\arguments{
\item{dv}{A data frame with the dependent variables}

\item{iv}{A data frame or vector with the independent variable}

\item{conditions}{A character vectot of length two with the names of the two conditions.
Defaults to the first two levels of the independent variable 'iv' if applicable.}

\item{labels}{A character vector of length two with labels for the dependent variables.}

\item{concise}{A more concise table with mean and sd in one column.}

\item{nice_p}{If TRUE, p values are printed in a nice format.}

\item{digits}{Number of digits for rounding mean and sd values}

\item{var_equal}{If FALSE, a t-test for unequal variances is calculated.}

\item{order}{Either "12" or "21" depicting whether group two is compared to group one or vice versa.}

\item{type}{Either "df" for data frame or "html" for html table.}
}
\value{
A tibble
}
\description{
Creats t.test tabel for multiple dependent variables
}
\examples{
dv <- data.frame(
  a = c(runif(85) * 150, runif(115) * 100), 
  b = runif(200) * 100
  )
iv <- factor(c(rep("A", 85), rep("B", 115)))
t_test_table(dv, iv, labels = c("Motivation", "Achievement"))
}
