% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{package}
\name{wmisc-package}
\alias{wmisc-package}
\alias{wmisc}
\title{Misc Functions}
\description{
A collection of functions and procedures
}
\author{
Juergen Wilbert [aut, cre]
}
\keyword{package}
